Refact UserApiV1RestControllerTest
